fix get client after refactoring
