+++ lib/creates/create_message.js
@@ -4,7 +4,7 @@
  * @copyright (c) 2019. Ingram Micro. All Rights Reserved.
  */
 
-const getConnectClient = require('../utils').getConnectClient;
+const { getConnectClient } = require('../connect/http');
 
 
 const createMessage = async (z, bundle) => {
